Separate LDFLAGS from LDADD/LIBADD
This affects the order of the linker command line in ways that matter
if you have a linker that defaults to --as-needed, like Ubuntu's, or
if you are linking statically. Dependency libraries (LDADD/LIBADD)
must appear last on the link line (so that they can satisfy
unresolved symbols of earlier items on the link line), while LDFLAGS
appear first.

Signed-off-by: Simon McVittie <smcv@collabora.com>