Cache the "real" implementation of dlsym at init time
The dlsym wrapper needs to cache the implementation of dlsym that's
in the GOT at the point when it (the wrapper) is installed.

Moving the implementation of the wrapper helper to capsule_init's
file so that capsule_init can set some static state to do this for
us.

This is enough to make openarena work.