set_elf_constraints: dlopen() is not safe inside a dlmopen() namespace
This can segfault, but depends on the exact artefacts that come out
of the compiler. dlmopen(LM_ID_BASE…) should be safe enough though.

In addition there's no point in walking the full code path again:
It isn't possible for the elf class & machine to change at runtime
so we might as well cache the results and return early for subsequent
calls.