Implement external dlopen wrapping
We need to intercept dlopen calls made by DSOs outside the capsule
so that we can update the GOTs of new DSOs as they appear (for
example when libSDL dlopen()s DSOs that use libGL's symbols but
don't link against it directly.

In order to do this we need to wrap dlopen(), but unlike dlsym()
(which we already wrap) we cannot do so unconditionally: To support
this we introduce a restricted relocate call which blacklists
certain DSOs.