capsule-shim.mk: Don't always update the symbols file
In automated build systems we won't be able to guarantee that the
sysroot will always be available at build time, and in any case we
shouldn't randomly change our ABI depending on what's in a sysroot
external to the source tree.

A timestamp difference between the symbols list and the shared
library list could be caused by more or fewer libraries being included,
but it could equally be caused by timestamp skew (git and most other
VCSs make no guarantees about timestamps), so don't require an update
unless the shared library list is actually different.

Signed-off-by: Simon McVittie <smcv@collabora.com>