Update library lists: treat more X libraries as core libraries
We certainly can't replace glibc, because libc and libdl are loaded
before libcapsule, and can have arbitrary internal state.

We might be able to encapsulate some of the X11 libraries later,
but for now they're a low priority: if they have to come from the host
system, that isn't terrible.

Signed-off-by: Simon McVittie <smcv@collabora.com>