tests/manual/gl.pl: Treat more X libraries as core libraries
We need to sync up libX11, libX11-xcb, libcapsule and libelf anyway,
so while we're experimenting with what works and what crashes,
we might as well exclude more of the X11/xcb family from encapsulation.

We might be able to take some or all of these out again when we have
everything working.

Signed-off-by: Simon McVittie <smcv@collabora.com>