tests/manual/gl.pl: Document how to make GLVND libGL work with older X11
For the practically useful case (GL provider = host), the X11 server
and libGL will likely be switched over to GLVND in lockstep, so the
only case where this is interesting is if the GL provider is not
actually the host. This is only going to happen during testing,
not in real life.

Signed-off-by: Simon McVittie <smcv@collabora.com>