process_pt_dynamic: Iterate over the dynamic section entries twice
In the first pass, we record the information we'll need in the second
pass. In the second pass, we do the actual work.

In particular this appears to fix a crash I was seeing with OpenArena
on a Debian 9 runtime on a Debian 8-based host, where the symbol
table appeared to be NULL - it looks as though that was crashing
because for that particular library, the DT_SYMTAB happened to appear
later in the dynamic section entry array than the DT_RELA or DT_JMPREL.

This should also let us stop using find_value(), but I haven't tried
that yet.

Signed-off-by: Simon McVittie <smcv@collabora.com>