process_pt_dynamic: Log if we see a DT_REL section
We relocate DT_REL entries in a DT_JMPREL section, and DT_RELA entries
either in a DT_JMPREL or DT_RELA section, so for symmetry it seems
like we should maybe be willing to relocate DT_REL entries in a
DT_REL section? For now, log it so we know whether it even happens.

Signed-off-by: Simon McVittie <smcv@collabora.com>