find_strtab: Redo calling convention with int/pointer the right way round
base is a valid pointer into memory, whereas start is only an offset
and so is not a pointer that we can dereference. Previously, base was
a pointer-sized integer and start was a pointer, which is exactly
backwards if we want to stick to a rule of "it's a pointer iff
you could dereference it".

Signed-off-by: Simon McVittie <smcv@collabora.com>