capsule-symbols: make this work with capsule shim libraries again
The way this tool loaded libraries was not compatible with the new
libcapsule infrastructure if the target was itself a capsule proxy:

The fix for this is to make capsule-symbols open the target into
the default namespace by calling ld_libs_load with LM_ID_BASE as
the requested namespace.

This does not affect capsule proxies, as they still pass LM_ID_NEWLM
to that function.