Always have exactly one link map per capsule_namespace
To preallocate data structures during metadata gathering, we need to
know ahead of time which capsules will be sharing a namespace. If we
need more flexibility later, we'll have to change from one
capsule_namespace per prefix to one capsule_namespace per
(prefix, discriminator) tuple, where the discriminator is something
from the metadata that identifies the desired equivalence classes of
libraries.

(In practice it's going to be easier to bind-mount the same prefix
into multiple places.)

Signed-off-by: Simon McVittie <smcv@collabora.com>