init: Drop all references to capsule metadata before unloading it
The capsule_namespace data structure contains "borrowed" pointers to
strings from all the libraries that share the namespace, and each
capsule_metadata data structure contains the same "borrowed" pointers.
Before we introduced capsule_namespace, this was equally true for the
tree_data. This is not safe if we can unload capsules.

Signed-off-by: Simon McVittie <smcv@collabora.com>