Include items to relocate in capsule_metadata
It turns out that all calls to capsule_relocate() wanted the same list
of items, but without being able to access the arguments of the
original capsule_relocate() call directly, capsule_external_dlopen()
had to rely on the items being saved ("cached") in the capsule struct.
By putting the items in the metadata, we can simplify that a lot.

Signed-off-by: Simon McVittie <smcv@collabora.com>