tests: use nm -B (like libtool does) to enumerate symbols
This protects us from potential bugs in capsule-symbols(1) by
having something else to cross-check our results against.
It also means we can inspect the generated shims for libraries that
we do not actually have and hence cannot load, such as libGL.so.1 on
autobuilders that do not have a graphics stack installed.

Signed-off-by: Simon McVittie <smcv@collabora.com>