Revert "init: tweak the list of libraries we never encapsulate"
In the case where we're running a Debian 9 container on a SteamOS
brewmaster (basically Debian 8) host, with glibc 2.24 in the container
and glibc 2.19 in the host, we have to make sure libpthread matches
libc. This is because libc 2.19 exported h_errno@@GLIBC_PRIVATE
which was imported by libpthread 2.19, but libc 2.24 no longer exports
that private symbol.

More generally, mixing versions of libraries from the same source
package is very hard to justify - if we reported bugs with this
configuration, upstream would be entirely justified in telling us that
we needed to fix our environment and ignoring the bug report.

This does reintroduce the deadlock in Debian 8 openarena, but it seems
that Debian 9 openarena (Debian 9 container, SteamOS host) survives
this configuration.

This reverts commit 096de1aa01a0c3be4faf4bfce281a96a36f207d5.

Signed-off-by: Simon McVittie <smcv@collabora.com>