tests/manual/gl.pl: Provide shortcuts for some plausible defaults
libcapsule seems to work OK for Mesa, so we can seriously consider
using it for that configuration.

With the NVIDIA binary driver, libcapsule crashes horribly; but the
benefit we can get from libcapsule is limited in that case anyway. The
NVIDIA driver is binary-only, so we are going to be using the precompiled
binary that matches the kernel, regardless of whether it was intended
for use with our container's base OS. However, we can assume that NVIDIA
compiled it with conservative glibc and libX11 dependencies, because
if they hadn't, it couldn't work on older OSs and containers anyway; so
we presumably don't really need to pull in newer supporting libraries,
just the NVIDIA blobs.

Signed-off-by: Simon McVittie <smcv@collabora.com>