ld_libs_load_cache: Don't reopen cache, and fix prefix="" case
We were opening the correct ld.so.cache here (in both cases), but then
calling ld_cache_open() which would close the fd we just opened and
open libs->prefix.path instead. In the case where the prefix is set
to "" we would open whatever arbitrary string happened to have been
left in libs->prefix.path last time we used it as scratch space.

Signed-off-by: Simon McVittie <smcv@collabora.com>