ld_libs_load_cache: Don't use prefix.path as scratch space
Previously, we'd write /etc/ld.so.cache into prefix.path[prefix.len:]:

    /host/etc/ld.so.cache\0\0...
    |<->||<------------>|
     len   temporarily used

and delete it again later with sanitise_ldlibs(). But that seems
unnecessarily confusing: the contents of prefix.path when interpreted
as an ordinary \0-terminated C string are sometimes the prefix, but
sometimes edited in-place to be the name of a file within the prefix.

Signed-off-by: Simon McVittie <smcv@collabora.com>