debian: Enable Gitlab-CI
We are targeting scout, but because scout is so old, its SDK environment
is too old to run deb-build-snapshot. Instead, we do that step in a
Debian 10 'buster' environment, and hand over the artifacts from that
to the autopkgtest step.

Our build-dependencies are not all preinstalled in the CI Docker
images, so add them.

Signed-off-by: Simon McVittie <smcv@collabora.com>