Add checks for locales and locale-related issues
Some Steam games assume that the en_US.UTF-8 locale is available, and
it's reasonably likely that others assume that the C.UTF-8 locale is
available, or that the locale environment variables are set to usable
values.

Having checks for locale properties will also help us when experimenting
with containers: we can check whether the locale that was set outside
the container is available inside the container.

Signed-off-by: Simon McVittie <smcv@collabora.com>