graphics: Only allow EGL_X11 with a GL-based rendering interface
EGL is specifically for OpenGL and OpenGL ES, so when we add Vulkan
(or, hypothetically, other rendering interfaces), it would make no
sense to request EGL_X11 in combination with a non-GL-based window
system. The only combination that makes sense for Vulkan in the short
term is (X11, VULKAN); if we later add support for non-X11 windowing
systems like Wayland, then (WAYLAND, VULKAN) would also make sense.

Signed-off-by: Simon McVittie <smcv@collabora.com>