WIP: capture-libs: Treat unversioned libraries as in indeterminate order
On at least Debian, Ubuntu and Manjaro, libgcc_s.so.1 is a regular
file, not a symlink to a versioned name (libgcc_s.so.1.2.3) like most
shared libraries. However, on Fedora 30 it is a symlink to a versioned
name like libgcc_s-9-20190827.so.1.

This can cause problems if a Debian-derived container is used on a Fedora
host, with host graphics drivers.

TODO: it isn't clear whether this change makes any practical difference,
because strverscmp("libgcc_s.so.1", "libgcc_s-9-20190827.so.1") == 0
anyway.