Detect graphics library vendor
With this change we are now able to check if the used driver is GLVND or
non-GLVND.
This information is useful for example when steam-runtime-system-info is
failing to find GLX ICDs. When this will happens we will be able to know
if this is due to the use of non-GLVND driver or if there is something
unexpected.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>