system-info: Find steamrt expectations relative to libsrt
If steam-runtime-system-info is run from an unpacked LD_LIBRARY_PATH
Steam Runtime, we want it to find that runtime's library expectations.
In particular, when using "steam-native" on Arch Linux, we want to
diagnose whether the system-wide libraries provided by Arch Linux are
compatible with the ones from the real Steam Runtime.

Signed-off-by: Simon McVittie <smcv@collabora.com>