capture-libs: Compare symbols and versions to determine order
When two libraries have the same numeric tail we were not able to
reliably determine which one was the newer.

Now we also check the libraries definitions and the symbols to make a
more weighted decision.

The order of comparison is currently:
compare version definitions > compare numeric tail > one has strictly
more symbols

If even after these three checks we are still not sure which one to
choose, we pick the provider as the default.