Revert "Build -system-info and -check-locale with relative DT_RPATH"
This made them usable from outside the LD_LIBRARY_PATH runtime, but breaks
their use inside a pressure-vessel container. In pressure-vessel, the
glibc family of libraries are typically taken from outside the container
(for example /overrides/lib/x86_64-linux-gnu/libc.so.6), and the dynamic
linker ld.so is replaced with one that is suitable for the libraries
in /overrides. In general, these versions will be incompatible with the
ones in /usr/lib/x86_64-linux-gnu, which unfortunately are the ones that
get pulled in via the DT_RPATH.

This reverts commit 91478a07add1b29eff0da03ae0667b10c34fb059.

Signed-off-by: Simon McVittie <smcv@collabora.com>