wrap: Mount pressure-vessel on /run/pressure-vessel if using runtime
When we're using a non-trivial runtime, we need to make
pressure-vessel-with-lock available to the container.

Previously, if your pressure-vessel installation was visible via your
home directory (if shared), via $(realpath ~/.steam/steam), or via the
current working directory at the time of running pressure-vessel-wrap,
then it would exist in the container at the same path as outside the
container, so this would work. However, if you have installed it in
an alternative Steam library such as /media/big/library, none of these
mount points would make it visible and bwrap would fail to run it.

Circumvent this by mounting the relocatable pressure-vessel prefix
on /run/pressure-vessel, which lets us run it via that path.

Bug: https://github.com/ValveSoftware/steam-runtime/issues/204
Signed-off-by: Simon McVittie <smcv@collabora.com>