wrap: Don't pass --lock-file to bwrap
This is redundant with us passing the fd to be held by
pressure-vessel-with-lock, and it causes bwrap to produce the wrong
exit status (see https://github.com/containers/bubblewrap/issues/336
and https://github.com/containers/bubblewrap/pull/325). It should
have been removed when the "pass the lock fd" code path was added.

Fixes: 959fd338 "wrap: Take out a lock on the container's runtime for the duration"
Signed-off-by: Simon McVittie <smcv@collabora.com>