tests: Check error indicator before checking boolean result
If a GLib function that raises an error fails unexpectedly during
testing, it's a lot more informative to see an assertion failure that
includes details of the error, rather than just the boolean result.

For a conventional GLib function we should not look at the "out"
parameters until we have checked that the function was successful.

Signed-off-by: Simon McVittie <smcv@collabora.com>