build: Consistently sort dependencies in stack order
It's slightly preferable to list linker and compiler flags in stack
order, with the lowest in the stack first. This allows use of a
higher-level library from a non-standard prefix without also necessarily
picking up lower-level libraries from the same non-standard-prefix.
For example, if /path/to/json-glib also includes a copy of GLib, then

    -L/path/to/glib -lglib-2.0 -L/path/to/json-glib -ljson-glib-1.0

will link the GLib from /path/to/glib, but

    -L/path/to/json-glib -ljson-glib-1.0 -L/path/to/glib -lglib-2.0

will link both the GLib and the json-glib from /path/to/json-glib.

Signed-off-by: Simon McVittie <smcv@collabora.com>