graphics: Search for extra drivers in libQUAL/dri, not in libQUAL
When we found a loader library in ${prefix}/libQUAL, we were
correctly looking in ${prefix}/libQUAL/dri.

However, when we found a Debian-style loader library in
${prefix}/lib/${multiarch}, we were correctly looking for preferred
drivers in ${prefix}/lib/${multiarch}/dri, but we were looking for extra
drivers in ${prefix}/lib and ${prefix}/libQUAL, not in ${prefix}/lib/dri
and ${prefix}/libQUAL/dri as we should.

Signed-off-by: Simon McVittie <smcv@collabora.com>