tests/capture-libs: Tolerate path to ld.so not being fully resolved
--resolve-ld.so would ideally be the equivalent of realpath or
readlink -f, but in fact it only has special handling for symlinks in
the last position in the pathname (like AT_SYMLINK_NOFOLLOW and
O_NOFOLLOW), leading to different results on at least Arch Linux.

The pressure-vessel utility for which this option was added does not
actually use it any more, so fixing this is not a high priority right now.

Signed-off-by: Simon McVittie <smcv@collabora.com>