CI: Codify how to build on Arch Linux
Some libcapsule users and contributors are using Arch Linux or Manjaro
rather than a Debian derivative.

Many of the tests will be skipped on Gitlab-CI because they need a
working bubblewrap, which isn't allowed inside unprivileged Docker;
but this provides "executable documentation" for how to do a build
and test.

Signed-off-by: Simon McVittie <smcv@collabora.com>