Log driver-selection environment variables
Some drivers have an environment variable that overrides the automatic
detection of which driver should be used. For example Mesa has
`MESA_LOADER_DRIVER_OVERRIDE`, VA-API has `LIBVA_DRIVER_NAME` and so on.

We now log these environment variables in the s-r-s-i output to have a
move complete view about the state of the running system.

Signed-off-by: Ludovico de Nittis <ludovico.denittis@collabora.com>