wrap: Remount API filesystems when operating without a runtime
In the case where we bind-mounted the entire root filesystem, since
commit 77f3e3d7 we mounted it over the top of API filesystems like
/dev, /proc, /sys. This results in /dev being mounted with the nodev
option, which breaks a lot of normal functionality like /dev/null.
Some games can work despite this, but only badly (Floating Point
launches with indirect or software rendering and a horrible framerate),
and the xterm activated by PRESSURE_VESSEL_SHELL=instead fails to launch
because it can't create a pseudo-terminal.

Adjust the order to get this right.

Fixes: 77f3e3d7 "wrap: Make sure the intermediate bwrap object can be used as a base"
Signed-off-by: Simon McVittie <smcv@collabora.com>