utils: Add a method to copy a directory tree using hard links
This will let us duplicate a runtime and edit it in-place. The major
appeal of doing this is that it's something we can do in a Flatpak
environment, where recursively invoking bubblewrap isn't allowed.

It also seems like it might yield a more reliable way to overwrite parts
of the runtime with their host-system equivalents than the tricks we
currently use with files and directories mounted over their runtime
counterparts.

The major down side is that after we've done this, we have a copy of
the runtime, which we need to garbage-collect and clean up eventually.

Signed-off-by: Simon McVittie <smcv@collabora.com>