runtime: Optionally copy the runtime, and bind-mount the copy
This is good for two reasons. In a Flatpak runtime, we will not be able
to do tricks with arbitrary bind-mounts; if we want the host library
stack, we will have to rewrite the mutable copy of the runtime to
replace container libraries with symlinks into /run/host. Also, when
not in a Flatpak runtime, having a mutable copy of the immutable runtime
gives developers a chance to edit the runtime to inject workarounds or
things they're testing.

Signed-off-by: Simon McVittie <smcv@collabora.com>