bwrap: Don't create a /libexec symlink
We want all the libQUAL directories, including lib32 and lib64 (and in
principle also libx32, libn32 etc.), but we specifically don't care
about libexec. Real systems don't usually symlink /usr/libexec into the
root directory.

If we create a /libexec symlink, the only effect it might have is that
game developers start relying on it, which we certainly don't want.

Signed-off-by: Simon McVittie <smcv@collabora.com>