runtime: Access runtime without using bwrap if possible
If the runtime is already "the right shape", then we don't need to
use bwrap to mount it on /usr and populate a suitable tmpfs root
filesystem.

We need to do this if we want to support being run from inside a
Flatpak, because we can't run bwrap inside Flatpak. It might also
speed up accesses to suitable runtimes, because we won't waste time
invoking extra bwrap processes.

Signed-off-by: Simon McVittie <smcv@collabora.com>