scripts: Assert that we have the desired Python version
In Steam Runtime 1 'scout' and Steam Runtime 1½ 'heavy' SDK
environments, we have a backport of python3.5 in order to run Meson,
and meson.build automatically selects it if available.

In newer environments like a developer's laptop, we expect to have a
strictly newer Python version, like the python3.7 in Debian 10.

One notable exception is that we only require Python 3.4+ for scripts
that are meant to run on test machines or on non-developer Steam users'
systems, which right now just means the test UI. This version was chosen
because SteamOS 2 'brewmaster' and Ubuntu 14.04 'trusty' both have 3.4.

Ubuntu 12.04 'precise' only has 3.2, but the Steam client doesn't
actually work on 12.04 any more for orthogonal reasons.

Signed-off-by: Simon McVittie <smcv@collabora.com>