populate-depot: New script to download one or more runtimes
As we start to add more runtimes, we need to decouple the
SteamLinuxRuntime depot from any individual runtime. Instead of producing
a pre-populated depot as part of a scout build, we can clone this git
repository in some convenient place and run ./populate-depot.py to
download the necessary files.

Also add convenience Makefile targets to download scout and optionally
soldier - these are mostly here as manual tests and as executable
examples.

This partially resolves T17590 and T20721.

Signed-off-by: Simon McVittie <smcv@collabora.com>