runtime: Create /overrides on disk if using a mutable sysroot
This means we're able to inspect it in --only-prepare mode.

At the moment we still create /overrides/bin/localedef,
/overrides/bin/locale, and sometimes EGL and Vulkan JSON via --symlink
or --ro-bind, so we have to make /overrides read/write from the
perspective of the container. However, when we have moved to creating
those files as symlinks in /overrides, we will be able to lock it down
to read-only from the perspective of the container.

Signed-off-by: Simon McVittie <smcv@collabora.com>