locale-gen: Add --output-dir option
At the moment it writes to the current working directory, and we invoke
it via `bwrap ... --chdir DIRECTORY`; but we want to make it possible to
use this code from inside a Flatpak runtime (at least in the case where
we're using the host system's glibc), which means we won't be allowed
to run bwrap and must do this some other way.

Rather than using `env --chdir=...` (which isn't supported in scout
anyway), let's teach this script to output to an explicitly specified
directory.

Signed-off-by: Simon McVittie <smcv@collabora.com>