runtime: Copy pressure-vessel itself into mutable runtime if needed
We only strictly need pressure-vessel-with-lock and GLib (and we could
skip even the GLib dependency if we assume the runtime contains
GLib >= 2.32, as all ours do), but the relocatable copy is designed to
be portable to any OS, so for now let's copy the whole thing in.
When we do it like this, it's in /usr, because we anticipate that
Flatpak sub-sandboxing is not going to give us control over /run.

Signed-off-by: Simon McVittie <smcv@collabora.com>