runtime: Remove dead code
The entire setup for this call to `readlink -f` should have been removed
when it was hoisted further up in bind_runtime() (and replaced with
`readlink -e`) during commit 26b3be64, but in fact only the part that
actually *ran* the wrapped readlink command was deleted.

This also always leaked some memory.

Signed-off-by: Simon McVittie <smcv@collabora.com>